Fix destructive find_object_to_grasp method
The find_object_to_grasp() method calls apply_planar_transformation() on
the input MHI, causing the image to change. In Rviz, it looks as if the
image has been shifted down a few inches. Running find_object_to_grasp()
over and over on the same MHI will eventually destroy the MHI to the
point where a surface or grasp target is unable to be found.