Experiments with segmenting graspable objects (#3)
* Add manip view test assets

* Cythonize in setup.py

* Add test segment object

* Add get_surface performance test

* Add second get_surface performance test

* Fix destructive find_object_to_grasp method

The find_object_to_grasp() method calls apply_planar_transformation() on
the input MHI, causing the image to change. In Rviz, it looks as if the
image has been shifted down a few inches. Running find_object_to_grasp()
over and over on the same MHI will eventually destroy the MHI to the
point where a surface or grasp target is unable to be found.

* Add get_object performance tests

* Started find_objects_to_grasp() method

* Completed find_objects_to_grasp() method

* Split performance tests for segment_objects

* Fix bug with yaml dumping

* Add functional test for find_objects_to_grasp()

* Add performance tests for find_objects_to_grasp()

* Add the ability substitute mhi in to_point_cloud()

* Add change-able floor TF to Navigate.forward()

* Clean whitespace in manipulation_planning.py

* Fix floor_tf always set to odom on backwards

* Add find_placement_target() method