Continue splitting navigate.py and ros_navigate.py
To create a Pyfunmap version of MoveBase.local_plan(), this commit
starts to port methods in FastSingleViewPlanner to Pyfunmap. The
ROS version of this class is moved to the ros_navigate.py module.
Supporting this port are changed in MHI and RMHI. Additionally,
the mapping modules are changed according to the new changes in
MHI/RMHI.